+++ unbound/web/config_gen.py
@@ -34,6 +34,10 @@ CONFIG_SCHEMA = {
         "type": "bool",
         "default": False,
     },
+    "allow_large_blocklists": {
+        "type": "bool",
+        "default": False,
+    },
     "access_control": {
         "type": "list",
         "default": ["127.0.0.0/8", "10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"],